requirements.txt
numpy==1.20.2
numba==0.51.2
importlib==1.0.4
matplotlib == 3.4.1
Pillow==8.2.0
palettable==3.3.0
pygame==2.0.1
# OpenGL
# colorcet